+++ src/index.ts
@@ -1,6 +1,12 @@
 export * from "./lib/DeleteAction/DeleteAction";
-export * from "./lib/Utils/DateHandler";
 export * from "./lib/Layout/AuthenticationLayout";
+
+// Panel sidebar layout
+export * from "./lib/Layout/PanelSideBarLayout/PanelSideBarLayout";
+export * from "./lib/Layout/PanelSideBarLayout/PanelSideBar/Context/PanelSideBarContext";
+export * from "./lib/Layout/PanelSideBarLayout/PanelSideBar/Definitions/PanelItem";
+
+// Sidebar layout
 export * from "./lib/Layout/SideBarLayout/SideBarLayout";
 export * from "./lib/Layout/SideBarLayout/SideBarLayoutContext";
 export * from "./lib/SideBar/SideBarMenuContext";