+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `PanelSidebar` is new sidebar with custom side panels and a new `PanelSideBarLayoutContext`
+- `collapseIconOnly` to `PanelItem` in order to collapse `PanelSideBarItem` only by clicking on chevron icon and label can work as link
+
+### Removed
+
+- :boom: `DateHandler` utils as it belongs to another package and will be replaced in the future
+
 ## [2.10.0] - 2023-10-20
 
 ### Added